+++ CHANGES.md
@@ -35,6 +35,8 @@ Changes 1.8.0-prerelease
 
 * ML-DSA: slicing the implementation to allow keygen, siggen and sigver to be compiled independently - together with dead code stripping, only the desired functionality is present (e.g. in secure boot mode, only sigver is of interest)
 
+* Linux kernel: add HMAC support
+
 Changes 1.7.2
 * Fix RDSEED counter